S20-579:refactor components
