+++ package.json
@@ -26,7 +26,6 @@
   },
   "husky": {
     "hooks": {
-      "pre-commit": "lint-staged"
     }
   },
   "dependencies": {
