+++ src/components/Checkbox/index.ts
@@ -1,4 +1,1 @@
-import { Checkbox, CheckboxProps } from "./Checkbox";
-
-export { Checkbox };
-export type { CheckboxProps };
+export * from "./Checkbox";
