+++ src/components/Circle/index.ts
@@ -1,0 +1,1 @@
+export * from "./Circle";
