+++ src/components/CirclePagination/CirclePagination.stories.tsx
@@ -4,15 +4,10 @@ import { withA11y } from "@storybook/addon-a11y";
 import { action } from "@storybook/addon-actions";
 import { Column, Spacer } from "../../styled";
 
-export default {
-  title: "CirclePagination",
-  component: CirclePagination,
-  decorators: [withA11y],
-};
-
-const dummyData = [1, 2, 3, 4, 5];
-
-export const CirclePaginationGroup: React.FunctionComponent = () => {
+/**
+ * <CirclePaginationGroup />
+ */
+const CirclePaginationGroup: React.FC = () => {
   const [activeItem, setActiveItem] = useState(1);
 
   const onClick = (index: number) => {
@@ -26,7 +21,19 @@ export const CirclePaginationGroup: React.FunctionComponent = () => {
         {activeItem}
       </Column>
       <Spacer mt={{ xxs: 3, lg: 5 }} />
-      <CirclePagination activeItem={activeItem} data={dummyData} onItemClick={(index: number) => onClick(index)} />
+      <CirclePagination
+        activeItem={activeItem}
+        data={[1, 2, 3, 4, 5]}
+        onItemClick={(index: number) => onClick(index)}
+      />
     </Column>
   );
+};
+
+export { CirclePaginationGroup };
+
+export default {
+  title: "CirclePagination",
+  component: CirclePagination,
+  decorators: [withA11y],
 };
