+++ src/components/CirclePagination/CirclePagination.style.ts
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+import { Box } from "../../styled/Box";
+
+/**
+ * <ButtonStyled />
+ */
+const ButtonStyled = styled(Box)`
+  padding: 0;
+  background: none;
+  border: none;
+  cursor: pointer;
+`;
+
+export { ButtonStyled };
