+++ src/components/Input/Input.stories.tsx
@@ -5,15 +5,12 @@ import { Spacer, Column } from "../../styled";
 import { Input } from "./Input";
 import { withA11y } from "@storybook/addon-a11y";
 
-export default {
-  title: "Input",
-  component: Input,
-  decorators: [withA11y],
-};
-
 const downloadIcon = <DownloadIcon height={25} width={25} />;
 
-export const Inputs: React.FunctionComponent = () => {
+/**
+ * <Inputs />
+ */
+const Inputs: React.FunctionComponent = () => {
   const [age, setAge] = useState<number | undefined | string>();
   const [name, setName] = useState<number | undefined | string>();
   const [email, setEmail] = useState<number | undefined | string>();
@@ -81,4 +78,12 @@ export const Inputs: React.FunctionComponent = () => {
       </form>
     </Column>
   );
+};
+
+export { Inputs };
+
+export default {
+  title: "Input",
+  component: Input,
+  decorators: [withA11y],
 };
