+++ src/components/Input/index.ts
@@ -1,5 +1,1 @@
-import { Input } from "./Input";
-import { InputProps } from "./InputTypes";
-
-export { Input };
-export type { InputProps };
+export * from "./Input";
