--- src/components/Select/dummyData.ts
@@ -1,5 +1,0 @@
-export const OPTIONS_DATA = [
-  { value: "chocolate", label: "Chocolate" },
-  { value: "strawberry", label: "Strawberry" },
-  { value: "vanilla", label: "Vanilla" },
-];
