+++ src/components/Select/index.ts
@@ -1,5 +1,5 @@
 import { SelectComponent } from "./Select";
-import { SelectComponentProps } from "./SelectTypes";
+import { SelectComponentProps } from "./Select.props";
 
 export { SelectComponent };
 export type { SelectComponentProps };
